requirements-dev.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.7
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.3.0
djangorestframework==3.11.0
Faker==4.1.1
futures==3.1.1
google-api-core==1.20.1
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
goslate==1.5.1
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
oauthlib==3.1.0
polib==1.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
